Extract hooks
