+++ src/hooks/chat.ts
@@ -4,6 +4,7 @@ import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import { io } from 'socket.io-client';
 import { accessTokenState } from '@/states/user/auth';
 import { RoomInfo, roomInfoState } from '@/states/roomInfoState';
+import { Friend } from '@/hooks/friends';
 import { useFetcher } from '@/hooks/fetcher';
 import {
   useDismiss,
@@ -13,6 +14,85 @@ import {
 import { toast } from 'react-toastify';
 
 export type RoomListType = 'all' | 'joined';
+
+interface SubscribeData {
+  sender: Friend;
+  payload: string;
+}
+
+interface Message {
+  room_msg_id: number;
+  room_id: string;
+  sender: Friend;
+  payload: string;
+  created: Date;
+}
+
+export function useChatMessages(roomId: string) {
+  const fetcher = useFetcher();
+  const data = useQuery<Message[]>({
+    queryKey: ['chat/room/messages', roomId],
+    queryFn: async () => {
+      const res = await fetcher(`/chat/room/${roomId}/messages`);
+
+      if (res.ok) {
+        const data = await res.json();
+        return data.messages;
+      }
+      return [];
+    },
+    refetchOnWindowFocus: false,
+  });
+
+  return data;
+}
+
+export function useChat(roomId: string) {
+  const queryClient = useQueryClient();
+  const socketRef = useSocketRef(`ws://${import.meta.env.VITE_BASE_URL}:9999`);
+  const { data: messages } = useChatMessages(roomId);
+
+  useEffect(() => {
+    const socket = socketRef.current;
+
+    const handleSub = (data: SubscribeData) => {
+      queryClient.setQueryData<Message[]>(
+        ['chat/room/messages', roomId],
+        (prevMsg) => {
+          const newMessage: Message = {
+            room_msg_id:
+              prevMsg && prevMsg.length
+                ? prevMsg[prevMsg.length - 1].room_msg_id + 1
+                : 1,
+            room_id: roomId,
+            sender: data.sender,
+            payload: data.payload,
+            created: new Date(),
+          };
+          return prevMsg ? [...prevMsg, newMessage] : [newMessage];
+        },
+      );
+    };
+
+    const handleAnnounce = (data: { payload: string }) => {
+      toast.info(data.payload);
+    };
+
+    socket.emit('join', { room: roomId });
+
+    socket.on('subscribe', handleSub);
+    socket.on('subscribe_self', handleSub);
+    socket.on('announcement', handleAnnounce);
+
+    return () => {
+      socket.off('subscribe', handleSub);
+      socket.off('subscribe_self', handleSub);
+      socket.off('announcement', handleAnnounce);
+    };
+  }, [roomId, queryClient, socketRef]);
+
+  return { messages, socket: socketRef.current };
+}
 
 export function useRoomList(type: RoomListType) {
   const queryClient = useQueryClient();
